Get counts for confocal scanner through counter logic
If the confocal scanner hardware module talks to the counter hardware module
directly to get the counts for a given pixel, we get concurrency issues with
other modules trying to access the counter through the counter logic.

In fact, only one logic module should ever talk to a hardware module (at least
for blocking stuff)!!!

This commit links the counter logic to the confocal scanner hardware module and
acquires the counts for each pixel by listening to the sigCounterUpdated signal.