Make confocal image save path prefix configurable.
Confocal images used to be saved in a folder 'Confocal'. As we want to have
different confocal scanners (Attocube/Thorlabs), the save location should be
configurable.

With this commit, you can set the 'save_prefix' key in the configuration of any
confocal GUI module to replace the default 'Confocal' folder name.